LayoutTests/accessibility/ax-thread-text-apis/text-marker-range-stale-node-crash-expected.txt

someContent
This tests that we create a text marker range from a stale node won't lead to crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.textMarkerRangeLength(markerRange) is 4
PASS successfullyParsed is true

TEST COMPLETE
